General: use folders for entities, use non-relative imports (order in some files), extract enrollments entities from helpTypes, adapt references
