+++ components/Layout/Navbar/Navbar.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { useQuery } from '@tanstack/react-query'
 import { AppShell, Divider, ScrollArea } from '@mantine/core'
 import NavbarItem from './NavbarItem'
 import {
@@ -14,9 +15,8 @@ import {
   IconUserCircle,
   IconUsers,
 } from '@tabler/icons-react'
-import { CurrentUserQueryOptions } from '../../../services/currentUser'
-import { ProjectSortAttribute, RequestState } from '../../../entities/ProjectInList'
-import { useQuery } from '@tanstack/react-query'
+import { CurrentUserQueryOptions } from '@/services/currentUser'
+import { ProjectSortAttribute, RequestState } from '@/entities/Project/ProjectInList'
 
 const mockAppVersion = 'v1.0.0'
