+++ components/User/UserList.tsx
@@ -5,7 +5,7 @@ import Theme from 'src/app/theme'
 
 import InfoMessage from '../Common/InfoMessage/InfoMessage'
 import SkeletonFull from '../Loader/SkeletonFull'
-import User from '@/entities/User'
+import User from '@/entities/User/User'
 import UserItem from './UserItem'
 
 interface UserListProps {
