+++ entities/Enrollment/EnrollmentChangeRole.ts
@@ -1,4 +1,4 @@
-import { ProjectRole } from '../Enrollment'
+import { ProjectRole } from './Enrollment'
 
 export interface EnrollmentChangeRole {
   role: ProjectRole
