+++ entities/User/UsersResult.ts
