+++ src/app/users/page.tsx
@@ -16,7 +16,7 @@ import Filter from '@/components/Common/Filter/Filter'
 import Interest from '@/entities/Interest'
 import UserFilterContent from '@/components/User/UserFilterContent'
 import UserList from '@/components/User/UserList'
-import { UserSortAttribute } from '@/entities/UserInList'
+import { UserSortAttribute } from '@/entities/User/UserInList'
 import Facility from '@/entities/Facility'
 import ResearchDepartment from '@/entities/ResearchDepartment'
 import Institution from '@/entities/Institution'
